[REFACT] UI 관련 프리팹과 스크립트 정리 및 코드 정리
UI 추가방법 :
- Player 캐릭터에게 PlayerUIController 스크립트 추가
- QuestList 프리팹과 Inventory 프리팹을 씬에 추가

UI 테스트 방법 :
- Q 키로 퀘스트 창을 엽니다.
- I 키로 인벤토리 창을 엽니다.
- U 키로 인벤토리에 테스트 아이템을 추가합니다.(각 칸에 2번씩 추가 가능)

현재 인벤토리 UI(PlayerUIController.myBag)가 제공하는 기능 :
- 아이템의 추가 : addItem()
- 선택된 아이템 참조 : getSelectedItem()
