--- ImGround/Assets/soungsoo/Inventory/EventSystem.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 224ed4eee98212b49b993d19a6b25cab
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
