--- ImGround/Assets/soungsoo/Inventory/Inventory.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 7777353d1eb578b48aada7353d028a9e
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
