--- ImGround/Assets/soungsoo/Inventory/UIScripts/PlayerUIController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 49de25996647c5c4e9356037831a47d5
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
