--- ImGround/Assets/soungsoo/Inventory/slot.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: cf72771f77279a64395a1ac816d0b8a6
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
