--- ImGround/Assets/soungsoo/Quest/MissionManager.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: b1ca91c68191a0b498b6fac465ed6f1c
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
