--- ImGround/Assets/soungsoo/Quest/Quest.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 6552945752ed185429633a39c0c5986e
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
