--- ImGround/Assets/soungsoo/Quest/QuestList.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 3ecefea349ae47a4ca7524912a1ef990
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
