+++ ImGround/Assets/soungsoo/UI/Inventory.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 7777353d1eb578b48aada7353d028a9e
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
