+++ ImGround/Assets/soungsoo/UI/UIScripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a6ceeb127a46ab54da3c6e3b2b8b7af4
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
